:shirt: Split methods into class
